[Puppi] move from double to float

+++ CommonTools/PileupAlgos/interface/PuppiAlgo.h
@@ -15,27 +15,27 @@ public:
   //Computing Mean and RMS
   void reset();
   void fixAlgoEtaBin(int i_eta);
-  void add(PuppiCandidate const &iParticle, double const iVal, uint const iAlgo);
+  void add(PuppiCandidate const &iParticle, float const iVal, uint const iAlgo);
   void computeMedRMS(uint const iAlgo);
   //Get the Weight
-  double compute(std::vector<double> const &iVals, double const iChi2) const;
-  std::vector<double> const &alphas() const { return fPups; }
+  float compute(std::vector<float> const &iVals, float const iChi2) const;
+  std::vector<float> const &alphas() const { return fPups; }
   //Helpers
   inline int etaBins() const { return fEtaMin.size(); }
-  inline double etaMin(int i) const { return fEtaMin[i]; }
-  inline double etaMax(int i) const { return fEtaMax[i]; }
-  inline double ptMin() const { return cur_PtMin; }
+  inline float etaMin(int i) const { return fEtaMin[i]; }
+  inline float etaMax(int i) const { return fEtaMax[i]; }
+  inline float ptMin() const { return cur_PtMin; }
 
   inline int numAlgos() const { return fNAlgos; }
   inline int algoId(uint iAlgo) const { return fAlgoId.at(iAlgo); }
   inline bool isCharged(uint iAlgo) const { return fCharged.at(iAlgo); }
-  inline double coneSize(uint iAlgo) const { return fConeSize.at(iAlgo); }
-  inline double neutralPt(int iNPV) const { return cur_NeutralPtMin + iNPV * cur_NeutralPtSlope; }
+  inline float coneSize(uint iAlgo) const { return fConeSize.at(iAlgo); }
+  inline float neutralPt(int iNPV) const { return cur_NeutralPtMin + iNPV * cur_NeutralPtSlope; }
 
-  inline double rms() const { return cur_RMS; }
-  inline double median() const { return cur_Med; }
+  inline float rms() const { return cur_RMS; }
+  inline float median() const { return cur_Med; }
 
-  inline double etaMaxExtrap() const { return fEtaMaxExtrap; }
+  inline float etaMaxExtrap() const { return fEtaMaxExtrap; }
 
 private:
   uint fNAlgos;
@@ -44,32 +44,31 @@ private:
   std::vector<double> fPtMin;
   std::vector<double> fNeutralPtMin;
   std::vector<double> fNeutralPtSlope;
-
   std::vector<double> fRMSEtaSF;
   std::vector<double> fMedEtaSF;
-  double fEtaMaxExtrap;
+  float fEtaMaxExtrap;
 
-  double cur_PtMin;
-  double cur_NeutralPtMin;
-  double cur_NeutralPtSlope;
-  double cur_RMS;
-  double cur_Med;
+  float cur_PtMin;
+  float cur_NeutralPtMin;
+  float cur_NeutralPtSlope;
+  float cur_RMS;
+  float cur_Med;
 
-  std::vector<double> fRMS;                          // raw RMS per algo
-  std::vector<double> fMedian;                       // raw Median per algo
-  std::vector<std::vector<double> > fRMS_perEta;     // final RMS used after eta corrections
-  std::vector<std::vector<double> > fMedian_perEta;  // final Med used after eta corrections
+  std::vector<float> fRMS;                          // raw RMS per algo
+  std::vector<float> fMedian;                       // raw Median per algo
+  std::vector<std::vector<float> > fRMS_perEta;     // final RMS used after eta corrections
+  std::vector<std::vector<float> > fMedian_perEta;  // final Med used after eta corrections
 
-  std::vector<double> fPups;
-  std::vector<double> fPupsPV;
+  std::vector<float> fPups;
+  std::vector<float> fPupsPV;
   std::vector<int> fAlgoId;
   std::vector<bool> fCharged;
   std::vector<bool> fAdjust;
   std::vector<int> fCombId;
-  std::vector<double> fConeSize;
-  std::vector<double> fRMSPtMin;
-  std::vector<double> fRMSScaleFactor;
-  std::vector<double> fMean;
+  std::vector<float> fConeSize;
+  std::vector<float> fRMSPtMin;
+  std::vector<float> fRMSScaleFactor;
+  std::vector<float> fMean;
   std::vector<int> fNCount;
 };
